all: Use go.sia.tech/core/rhp
