+++ autopilot/scanner.go
@@ -7,10 +7,10 @@ import (
 	"sync/atomic"
 	"time"
 
+	rhpv2 "go.sia.tech/core/rhp/v2"
 	"go.sia.tech/core/types"
 	"go.sia.tech/renterd/api"
 	"go.sia.tech/renterd/hostdb"
-	rhpv2 "go.sia.tech/renterd/rhp/v2"
 	"go.uber.org/zap"
 )
