+++ internal/stores/contracts.go
@@ -4,9 +4,9 @@ import (
 	"context"
 	"errors"
 
+	rhpv2 "go.sia.tech/core/rhp/v2"
 	"go.sia.tech/core/types"
 	"go.sia.tech/renterd/api"
-	rhpv2 "go.sia.tech/renterd/rhp/v2"
 	"gorm.io/gorm"
 )
