+++ internal/stores/contracts_test.go
@@ -7,10 +7,10 @@ import (
 	"reflect"
 	"testing"
 
+	rhpv2 "go.sia.tech/core/rhp/v2"
 	"go.sia.tech/core/types"
 	"go.sia.tech/renterd/api"
 	"go.sia.tech/renterd/hostdb"
-	rhpv2 "go.sia.tech/renterd/rhp/v2"
 	"gorm.io/gorm/schema"
 	"lukechampine.com/frand"
 )
