+++ internal/stores/types.go
@@ -7,8 +7,8 @@ import (
 	"fmt"
 	"math/big"
 
+	rhpv2 "go.sia.tech/core/rhp/v2"
 	"go.sia.tech/core/types"
-	rhpv2 "go.sia.tech/renterd/rhp/v2"
 )
 
 var zeroCurrency = currency(types.ZeroCurrency)
