+++ internal/testing/gouging_test.go
@@ -8,8 +8,8 @@ import (
 	"testing"
 	"time"
 
+	rhpv2 "go.sia.tech/core/rhp/v2"
 	"go.sia.tech/core/types"
-	rhpv2 "go.sia.tech/renterd/rhp/v2"
 	"go.sia.tech/siad/node/api/client"
 	stypes "go.sia.tech/siad/types"
 	"lukechampine.com/frand"
