+++ object/slab.go
@@ -5,8 +5,8 @@ import (
 	"io"
 
 	"github.com/klauspost/reedsolomon"
+	rhpv2 "go.sia.tech/core/rhp/v2"
 	"go.sia.tech/core/types"
-	rhpv2 "go.sia.tech/renterd/rhp/v2"
 	"golang.org/x/crypto/chacha20"
 )
